+++ ergo_std/src/cache/schema/delete_unused_diagnostic_sources.sql
@@ -1,0 +1,3 @@
+-- Clean up unused diagnostic sources.
+-- () -> ()
+DELETE FROM associated_diagnostic_sources WHERE value_u8 IS NULL;