+++ ergo_std/src/cache/schema/write_path.sql
@@ -1,0 +1,6 @@
+-- Write an associated path.
+-- (value[2], path) -> ()
+INSERT INTO associated_paths (
+	value_u8, value_l8,
+	path
+) VALUES (?,?,?);