+++ ergo_std/src/cache/schema/write_value.sql
@@ -1,6 +1,6 @@
 -- Write a stored value.
 -- (id[2], evaluated[2], type, data) -> ()
-INSERT OR REPLACE INTO stored (
+INSERT OR IGNORE INTO stored (
 	id_u8, id_l8,
 	evaluated_u8, evaluated_l8,
 	type,